+++ src/components/Footer/LegoLogo.jsx
@@ -1,35 +1,39 @@
+import React from "react";
+import styles from "./LegoLogo.module.scss";
+
 export function LegoLogo() {
     return (
       <>
-        <svg
-          viewBox="0 0 180 180"
-          fill="currentColor"
-          xmlns="http://www.w3.org/2000/svg"
-        >
-          <path d="M1.223 1.204V178.8H178.82l-.001-177.596H1.223Z" fill="#fff" />
-          <path
-            fill-rule="evenodd"
-            clip-rule="evenodd"
-            d="M1.223 178.8V1.204h177.595l.001 177.596H1.223Zm162.171-86.743c1.555-4.606 3.316-11.5 3.606-17.033.221-4.174-.341-10.02-4.278-14.174-3.141-3.319-7.94-4.994-14.264-4.994-7.989 0-14.426 2.82-19.136 8.397l-1.243 1.466-.826-1.734c-.674-1.413-1.156-1.992-1.89-2.82-3.219-3.642-8.617-5.488-16.044-5.487-6.804 0-12.95 2.039-17.308 5.742l-1.103.937-.76-1.237c-2.058-3.345-7.188-5.51-13.065-5.51-8.456.002-16.988 1.15-22.135 9.447l-2.133 3.424-.167-4.033c-.13-3.12-.415-4.416-1.866-5.952-1.596-1.693-4.282-2.516-8.21-2.513-5.768 0-10.103 2.515-13.256 7.694-4.853 7.752-16.583 33.25-16.324 45.369.17 7.865 5.418 12.679 14.043 12.878 7.108.172 12.758-1.652 16.347-5.274l.916-.926.88.952c3.226 3.502 8.13 5.432 13.811 5.432 7.234 0 13.42-2.248 16.973-6.171l.97-1.069.912 1.121c3.326 4.086 8.466 6.228 14.863 6.202 8.343-.037 16.135-3.871 21.376-10.512l1.374-1.741.767 2.083c3.23 8.783 11.984 10.092 16.921 10.095 16.765 0 24.005-11.503 30.249-30.059Z"
-            fill="#FFED00"
-          />
-          <path
-            fill-rule="evenodd"
-            clip-rule="evenodd"
-            d="M1.223 178.8V1.204h177.595l.001 177.596H1.223Zm163.283-73.742c2.321-4.784 6.688-16.757 7.495-23.222l.009-.071c.984-7.892 1.915-15.36-3.41-22.781-3.786-5.274-10.237-8.813-20.052-8.814-7.419 0-14.045 2.01-19.343 6.416-4.537-4.49-11.044-6.528-19.802-6.526-6.853 0-12.605 1.567-17.614 4.574-3.636-2.889-7.904-4.694-14.704-4.694-6.164-.001-14.818.32-21.347 5.863-1.509-3.079-4.577-5.497-13.04-5.497-7.796 0-14.156 3.688-18.327 10.535-5.569 8.895-17.39 34.582-17.088 48.614.225 10.563 8.146 17.84 19.62 18.113 6.895.161 12.767-1.315 17.248-4.289 4.052 2.92 9.23 4.497 14.984 4.497 6.886.001 13.098-1.807 17.704-5.053 4.173 3.308 9.648 5.096 15.842 5.096 8.036.003 15.503-2.772 21.535-8.051 4.114 5.246 11.162 7.973 19.721 7.974 16.023.001 24.346-9.854 30.569-22.684Z"
-            fill="#E3000B"
-          />
-          <path
-            fill-rule="evenodd"
-            clip-rule="evenodd"
-            d="M0 180V0h180l-.003 179.997L0 180Zm177.52-2.479.002-175.043H2.476L2.48 177.52h175.04Z"
-          />
-          <path
-            fill-rule="evenodd"
-            clip-rule="evenodd"
-            d="M167.615 54.301a4.305 4.305 0 0 1 4.309-4.313 4.307 4.307 0 0 1 4.315 4.313 4.304 4.304 0 0 1-4.315 4.312 4.302 4.302 0 0 1-4.309-4.312Zm4.311-3.457c1.918.001 3.477 1.545 3.472 3.463.005 1.918-1.56 3.467-3.472 3.467a3.458 3.458 0 0 1-3.453-3.467 3.448 3.448 0 0 1 3.453-3.463Zm1.679 4.81.515.9-1.117.007-.357-.73a3.643 3.643 0 0 0-.609-.874c-.176-.183-.337-.229-.734-.229l-.164.001-.002 1.83h-.921v-4.467h2.225c.959 0 1.409.535 1.409 1.243.002.675-.454 1.16-1.208 1.25l.002.025c.383.133.51.258.961 1.044Zm-2.469-2.868v1.227h.933c.624 0 .83-.304.821-.62 0-.394-.313-.61-.921-.61l-.833.003Zm-42.77 10.664c4.348-5.15 10.782-8.839 20.086-8.839 16.105 0 20.318 10.53 19.79 20.48-.313 5.98-2.232 13.104-3.666 17.365-6.37 18.931-14.02 30.902-31.43 30.9-7.945.002-15.216-3.098-18.081-10.908-5.233 6.622-13.231 10.944-22.35 10.989-6.97.034-12.39-2.429-15.835-6.66-4.09 4.514-11.011 6.581-17.89 6.583-6.18-.004-11.34-2.157-14.727-5.835-3.774 3.806-9.704 5.826-17.257 5.649-9.675-.232-15.08-6.037-15.258-14.095-.266-12.474 11.614-38.226 16.511-46.045 3.38-5.547 8.075-8.27 14.31-8.273 3.423.003 6.938.598 9.111 2.9 1.787 1.89 2.074 3.57 2.206 6.75 5.48-8.83 14.716-10.03 23.19-10.028 6.484 0 11.866 2.427 14.124 6.096 4.346-3.694 10.605-6.038 18.113-6.038 7.685.001 13.41 1.876 16.974 5.908.786.885 1.345 1.562 2.079 3.102ZM43.59 106.915c.286-3.016-1.973-6.273-10.986-4.709.895-2.005 2.226-4.753 3.737-7.873v-.001C41.183 84.334 47.874 70.519 48 65.114c.068-2.665-1.006-4.887-5.328-4.885-4.566-.002-7.43 1.908-9.776 5.762-5.088 8.125-15.893 32.635-15.666 43.229.128 6.056 4.823 8.446 10.04 8.568 7.19.167 15.472-1.852 16.321-10.874Zm18.35-13.03c-.6 1.614-1.728 4.994-2.639 8.571 2.958-.738 5.173-1.249 8.97-1.161 4.329.106 7.099 1.898 7.097 5.48 0 8.674-9.593 11.222-16.24 11.224-7.306 0-13.724-4.155-13.724-12.144 0-9.366 5.076-23.568 9.84-33.023 5.848-11.616 11.827-13.159 22.12-13.159 4.526 0 9.736 1.932 9.736 6.204 0 5.922-5.008 8.18-9.983 8.466-2.127.122-5.394.239-7.348.099 0 0-1.661 2.525-3.416 7.015 9.208-1.296 13.116.795 11.562 6.12-2.104 7.199-8.342 7.674-15.976 6.307Zm39.964-19.114c1.215-1.703 2.84-3.28 5.333-3.28 3.055 0 2.759 1.616 2.034 3.681-2.059 5.88 4.159 6.035 5.677 5.915 5.428-.423 8.419-3.93 9.155-9.904 1.008-8.134-5.845-11.257-14.532-11.257-14.48 0-20.164 8.856-25.586 21.516-2.547 5.947-5.755 16.11-5.754 22.582-.001 9.09 5.53 14.019 14.444 14.019 13.017 0 22.215-10.38 24.843-23.313.809-3.98 1.806-11.449-8.304-11.033-5.132.212-8.228 1.344-9.86 6.588-1.68 5.402 4.413 5.549 4.413 5.549-1.226 5.585-4.815 8.971-8.013 8.974-1.996 0-3.812-.866-3.204-5.09.903-6.176 6.687-21.198 9.354-24.947Zm60.017 7.407c-1.536 7.62-4.976 16.64-8.714 23.361-6.1 10.964-13.499 12.5-19.968 12.418-6.465-.077-13.754-2.461-13.809-12.439-.039-7.16 3.048-17.272 5.671-23.893 4.575-12.01 9.241-21.805 24.01-21.63 17.218.201 14.147 15.516 12.81 22.183Zm-21.598 19.537c2.394-4.055 9.133-22.448 9.317-27.047.054-1.34-.167-2.9-2.327-2.925-1.481-.021-2.769.266-3.927 2.055-2.602 3.382-10.219 24.185-10.09 28.335.047 1.492.88 2.795 2.712 2.795 2.11.002 3.279-1.45 4.315-3.213Z"
-          />
-        </svg>
+          <svg
+            className={styles.LegoLogo}
+            viewBox="0 0 180 180"
+            fill="currentColor"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <path d="M1.223 1.204V178.8H178.82l-.001-177.596H1.223Z" fill="#fff" />
+            <path
+              fill-rule="evenodd"
+              clip-rule="evenodd"
+              d="M1.223 178.8V1.204h177.595l.001 177.596H1.223Zm162.171-86.743c1.555-4.606 3.316-11.5 3.606-17.033.221-4.174-.341-10.02-4.278-14.174-3.141-3.319-7.94-4.994-14.264-4.994-7.989 0-14.426 2.82-19.136 8.397l-1.243 1.466-.826-1.734c-.674-1.413-1.156-1.992-1.89-2.82-3.219-3.642-8.617-5.488-16.044-5.487-6.804 0-12.95 2.039-17.308 5.742l-1.103.937-.76-1.237c-2.058-3.345-7.188-5.51-13.065-5.51-8.456.002-16.988 1.15-22.135 9.447l-2.133 3.424-.167-4.033c-.13-3.12-.415-4.416-1.866-5.952-1.596-1.693-4.282-2.516-8.21-2.513-5.768 0-10.103 2.515-13.256 7.694-4.853 7.752-16.583 33.25-16.324 45.369.17 7.865 5.418 12.679 14.043 12.878 7.108.172 12.758-1.652 16.347-5.274l.916-.926.88.952c3.226 3.502 8.13 5.432 13.811 5.432 7.234 0 13.42-2.248 16.973-6.171l.97-1.069.912 1.121c3.326 4.086 8.466 6.228 14.863 6.202 8.343-.037 16.135-3.871 21.376-10.512l1.374-1.741.767 2.083c3.23 8.783 11.984 10.092 16.921 10.095 16.765 0 24.005-11.503 30.249-30.059Z"
+              fill="#FFED00"
+            />
+            <path
+              fill-rule="evenodd"
+              clip-rule="evenodd"
+              d="M1.223 178.8V1.204h177.595l.001 177.596H1.223Zm163.283-73.742c2.321-4.784 6.688-16.757 7.495-23.222l.009-.071c.984-7.892 1.915-15.36-3.41-22.781-3.786-5.274-10.237-8.813-20.052-8.814-7.419 0-14.045 2.01-19.343 6.416-4.537-4.49-11.044-6.528-19.802-6.526-6.853 0-12.605 1.567-17.614 4.574-3.636-2.889-7.904-4.694-14.704-4.694-6.164-.001-14.818.32-21.347 5.863-1.509-3.079-4.577-5.497-13.04-5.497-7.796 0-14.156 3.688-18.327 10.535-5.569 8.895-17.39 34.582-17.088 48.614.225 10.563 8.146 17.84 19.62 18.113 6.895.161 12.767-1.315 17.248-4.289 4.052 2.92 9.23 4.497 14.984 4.497 6.886.001 13.098-1.807 17.704-5.053 4.173 3.308 9.648 5.096 15.842 5.096 8.036.003 15.503-2.772 21.535-8.051 4.114 5.246 11.162 7.973 19.721 7.974 16.023.001 24.346-9.854 30.569-22.684Z"
+              fill="#E3000B"
+            />
+            <path
+              fill-rule="evenodd"
+              clip-rule="evenodd"
+              d="M0 180V0h180l-.003 179.997L0 180Zm177.52-2.479.002-175.043H2.476L2.48 177.52h175.04Z"
+            />
+            <path
+              fill-rule="evenodd"
+              clip-rule="evenodd"
+              d="M167.615 54.301a4.305 4.305 0 0 1 4.309-4.313 4.307 4.307 0 0 1 4.315 4.313 4.304 4.304 0 0 1-4.315 4.312 4.302 4.302 0 0 1-4.309-4.312Zm4.311-3.457c1.918.001 3.477 1.545 3.472 3.463.005 1.918-1.56 3.467-3.472 3.467a3.458 3.458 0 0 1-3.453-3.467 3.448 3.448 0 0 1 3.453-3.463Zm1.679 4.81.515.9-1.117.007-.357-.73a3.643 3.643 0 0 0-.609-.874c-.176-.183-.337-.229-.734-.229l-.164.001-.002 1.83h-.921v-4.467h2.225c.959 0 1.409.535 1.409 1.243.002.675-.454 1.16-1.208 1.25l.002.025c.383.133.51.258.961 1.044Zm-2.469-2.868v1.227h.933c.624 0 .83-.304.821-.62 0-.394-.313-.61-.921-.61l-.833.003Zm-42.77 10.664c4.348-5.15 10.782-8.839 20.086-8.839 16.105 0 20.318 10.53 19.79 20.48-.313 5.98-2.232 13.104-3.666 17.365-6.37 18.931-14.02 30.902-31.43 30.9-7.945.002-15.216-3.098-18.081-10.908-5.233 6.622-13.231 10.944-22.35 10.989-6.97.034-12.39-2.429-15.835-6.66-4.09 4.514-11.011 6.581-17.89 6.583-6.18-.004-11.34-2.157-14.727-5.835-3.774 3.806-9.704 5.826-17.257 5.649-9.675-.232-15.08-6.037-15.258-14.095-.266-12.474 11.614-38.226 16.511-46.045 3.38-5.547 8.075-8.27 14.31-8.273 3.423.003 6.938.598 9.111 2.9 1.787 1.89 2.074 3.57 2.206 6.75 5.48-8.83 14.716-10.03 23.19-10.028 6.484 0 11.866 2.427 14.124 6.096 4.346-3.694 10.605-6.038 18.113-6.038 7.685.001 13.41 1.876 16.974 5.908.786.885 1.345 1.562 2.079 3.102ZM43.59 106.915c.286-3.016-1.973-6.273-10.986-4.709.895-2.005 2.226-4.753 3.737-7.873v-.001C41.183 84.334 47.874 70.519 48 65.114c.068-2.665-1.006-4.887-5.328-4.885-4.566-.002-7.43 1.908-9.776 5.762-5.088 8.125-15.893 32.635-15.666 43.229.128 6.056 4.823 8.446 10.04 8.568 7.19.167 15.472-1.852 16.321-10.874Zm18.35-13.03c-.6 1.614-1.728 4.994-2.639 8.571 2.958-.738 5.173-1.249 8.97-1.161 4.329.106 7.099 1.898 7.097 5.48 0 8.674-9.593 11.222-16.24 11.224-7.306 0-13.724-4.155-13.724-12.144 0-9.366 5.076-23.568 9.84-33.023 5.848-11.616 11.827-13.159 22.12-13.159 4.526 0 9.736 1.932 9.736 6.204 0 5.922-5.008 8.18-9.983 8.466-2.127.122-5.394.239-7.348.099 0 0-1.661 2.525-3.416 7.015 9.208-1.296 13.116.795 11.562 6.12-2.104 7.199-8.342 7.674-15.976 6.307Zm39.964-19.114c1.215-1.703 2.84-3.28 5.333-3.28 3.055 0 2.759 1.616 2.034 3.681-2.059 5.88 4.159 6.035 5.677 5.915 5.428-.423 8.419-3.93 9.155-9.904 1.008-8.134-5.845-11.257-14.532-11.257-14.48 0-20.164 8.856-25.586 21.516-2.547 5.947-5.755 16.11-5.754 22.582-.001 9.09 5.53 14.019 14.444 14.019 13.017 0 22.215-10.38 24.843-23.313.809-3.98 1.806-11.449-8.304-11.033-5.132.212-8.228 1.344-9.86 6.588-1.68 5.402 4.413 5.549 4.413 5.549-1.226 5.585-4.815 8.971-8.013 8.974-1.996 0-3.812-.866-3.204-5.09.903-6.176 6.687-21.198 9.354-24.947Zm60.017 7.407c-1.536 7.62-4.976 16.64-8.714 23.361-6.1 10.964-13.499 12.5-19.968 12.418-6.465-.077-13.754-2.461-13.809-12.439-.039-7.16 3.048-17.272 5.671-23.893 4.575-12.01 9.241-21.805 24.01-21.63 17.218.201 14.147 15.516 12.81 22.183Zm-21.598 19.537c2.394-4.055 9.133-22.448 9.317-27.047.054-1.34-.167-2.9-2.327-2.925-1.481-.021-2.769.266-3.927 2.055-2.602 3.382-10.219 24.185-10.09 28.335.047 1.492.88 2.795 2.712 2.795 2.11.002 3.279-1.45 4.315-3.213Z"
+            />
+          </svg>
       </>
     );
   }
