+++ src/components/Footer/LegoLogo.module.scss
@@ -1,0 +1,4 @@
+.LegoLogo{
+    width: 97px;
+    height: 97px;
+}
