+++ src/pages/CheckoutPage.module.scss
@@ -1,32 +1,48 @@
-.checkoutPagePayContainer {
-  margin: 1rem;
-  padding: 1rem;
-  width: auto;
+.checkoutPageGeneral{
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  margin: 0 auto;
+  max-width: 82.5rem;
 
-    .checkoutPagePayBtn {
-      background-color: #006db7;
-      border-radius: 9999px;
-      padding: 8px 15px;
-      align-self: center;
-    }
-  }
+   .checkoutPageProductsContainer {
+    margin: 1rem 1rem 1rem 0;
+    padding: 1rem 1rem 1rem 0;
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
 
-.checkoutPageProductsContainer {
-  margin: 1rem;
-  padding: 1rem;
-
-  .cardsProductsCont {
-    .cardsProductsList {
-      display: flex;
-      flex-direction: column;
-      gap: 1rem;
-      .cardProductCont {
+    .cardsProductsCont {
+      .cardsProductsList {
         display: flex;
-        flex-direction: row;
-        .productImg {
-          width: 10rem;
+        flex-direction: column;
+        gap: 1rem;
+        .cardProductCont {
+          display: flex;
+          flex-direction: row;
+          gap:1rem;
+          .productImg {
+            width: 10rem;
+          }
         }
       }
     }
   }
+
+  .checkoutPagePayContainer {
+    display: flex;
+    flex-direction: column;
+    justify-content: end;
+    gap: 1rem;
+    margin: 1rem;
+    padding: 1rem;
+    width: auto;
+
+      .checkoutPagePayBtn {
+        background-color: #006db7;
+        border-radius: 9999px;
+        padding: 8px 15px;
+        align-self: center;
+      }
+    }
 }
