+++ src/pages/PaymentResultCancel.jsx
@@ -1,3 +1,4 @@
+import { Link } from "react-router-dom";
 import Tipografia from "../components/Tipografia/Tipografia";
 import { useStoreState, useStoreActions } from "../store";
 import styles from "./PaymentResult.module.scss";
@@ -16,9 +17,14 @@ export const PaymentResultCancel = () => {
   };
 
   return (
-    <>
+     <div className={styles.payGeneral}>
       <div className={styles.payResultCont}>
-        <Tipografia texto="Ha ocurrido un error." color={"--red"} isTitleLarge isBoldWeight />
+        <Tipografia
+          texto="Ha ocurrido un error..."
+          color={"--red"}
+          isTitleLarge
+          isBoldWeight
+        />
 
         <img
           className={styles.payResultImg}
@@ -26,6 +32,18 @@ export const PaymentResultCancel = () => {
           alt="Lego icon happy"
         />
       </div>
-    </>
+      <div>
+        <button className={styles.checkoutPagePayBtn}>
+          <Link to="/">
+            <Tipografia
+              color={"--white"}
+              texto={"Volver a la página principal"}
+              isSubtitleRegular
+              isMediumWeight
+            />
+          </Link>
+        </button>
+      </div>
+    </div>
   );
 };
